Fixed seven busted tests, spotted by Alexey Dejneka:
--------------

Hello,

Thank you for your tests. They have shown two new bugs in the SBCL
compiler (in addition to >160 failed tests).

I think the following tests are wrong:

map-string.30: (SIMPLE-VECTOR size), not (SIMPLE-VECTOR type)

map-array.7: ARRAY is not a subtype of SEQUENCE

map-array.8: SIMPLE-ARRAY is not a subtype of SEQUENCE

map-string.31: --"--

map-string.32: (SIMPLE-ARRAY CHARACTER 5) is a 5-dimensional array

find-error.4, position-error.4: see CLHS 1.4.2 for 'should be prepared
to signal an error'

-- Regards, Alexey Dejneka


git-svn-id: svn://common-lisp.net/project/ansi-test/svn/trunk/ansi-tests@90 9bf3d7d6-7dd8-11dd-ad1d-697a6521c0a0